requirements.txt
flask
flask-cors
pyodbc
python-dotenv
gunicorn
